# docs/openapi/b1.yaml
openapi: 3.1.0
info:
  title: LapanganGo Phase B2 API
  version: 2.0.0
  description: Kontrak kumulatif B1 dan B2 untuk runtime lokal sandbox LapanganGo.
...
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
paths:
...
  /venues/{venueId}/reviews:
    get:
      tags: [Reviews]
      operationId: listVenueReviews
      parameters:
        - {
            name: venueId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      responses:
        "200":
          {
            description: Visible venue reviews,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
  /bookings/{bookingId}/review:
    post:
      tags: [Reviews]
      operationId: createBookingReview
      parameters:
        - {
            name: bookingId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/BookingReference" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201":
          {
            description: Review created,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /reviews/{reviewId}:
    patch:
      tags: [Reviews]
      operationId: updateCustomerReview
      parameters:
        - {
            name: reviewId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Review updated within seven-day window. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /reviews/{reviewId}/report:
    post:
      tags: [Reviews]
      operationId: reportCustomerReview
      parameters:
        - {
            name: reviewId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Review report recorded. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /business/reviews:
    get:
      tags: [Reviews]
      operationId: listBusinessReviews
      parameters:
        - {
            name: tenantId,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      responses:
        "200":
          {
            description: Tenant review management,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
  /business/reviews/{reviewId}/reply:
    post:
      tags: [Reviews]
      operationId: replyBusinessReview
      parameters:
        - {
            name: reviewId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201": { description: Owner reply recorded. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
  /admin/reviews/{reviewId}/moderate:
    post:
      tags: [Reviews]
      operationId: moderateAdminReview
      parameters:
        - {
            name: reviewId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Review visibility updated with reason. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /support/tickets:
    get:
      tags: [Support]
      operationId: listCustomerSupportTickets
      responses:
        "200":
          {
            description: Customer tickets,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
    post:
      tags: [Support]
      operationId: createCustomerSupportTicket
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201":
          {
            description: Ticket created,
            content:
              {
                application/json: { schema: { $ref: "#/components/schemas/B2Object" } },
              },
          }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /support/tickets/{ticketCode}/messages:
    get:
      tags: [Support]
      operationId: listCustomerSupportMessages
      parameters:
        - { name: ticketCode, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          {
            description: Customer-owned ticket thread,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
    post:
      tags: [Support]
      operationId: createCustomerSupportMessage
      parameters:
        - { name: ticketCode, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201": { description: Customer message recorded. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /business/support:
    get:
      tags: [Support]
      operationId: listBusinessSupportTickets
      parameters:
        - {
            name: tenantId,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      responses:
        "200":
          {
            description: Tenant support tickets,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
  /business/support/{ticketId}/messages:
    get:
      tags: [Support]
      operationId: listBusinessSupportMessages
      parameters:
        - {
            name: ticketId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
        - {
            name: tenantId,
            in: query,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      responses:
        "200":
          {
            description: Tenant-scoped ticket thread,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
    post:
      tags: [Support]
      operationId: createBusinessSupportMessage
      parameters:
        - {
            name: ticketId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201": { description: Tenant message recorded. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
...
  /admin/support/{ticketId}:
    patch:
      tags: [Support]
      operationId: updateAdminSupportTicket
      parameters:
        - {
            name: ticketId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "204": { description: Assignment, status, or resolution updated. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
  /admin/support/{ticketId}/messages:
    get:
      tags: [Support]
      operationId: listAdminSupportMessages
      parameters:
        - {
            name: ticketId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      responses:
        "200":
          {
            description: Platform ticket thread,
            content:
              { application/json: { schema: { $ref: "#/components/schemas/B2List" } } },
          }
    post:
      tags: [Support]
      operationId: createAdminSupportMessage
      parameters:
        - {
            name: ticketId,
            in: path,
            required: true,
            schema: { $ref: "#/components/schemas/PublicId" },
          }
      requestBody:
        required: true
        content:
          { application/json: { schema: { $ref: "#/components/schemas/B2Object" } } }
      responses:
        "201": { description: Admin message recorded. }
        "413": { $ref: "#/components/responses/PayloadTooLarge" }
components:
...
  responses:
...
    PayloadTooLarge:
      description: JSON request body exceeds the global 1 MB limit.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ApiError" }
...
  schemas:
    B2Object:
      type: object
      additionalProperties: true
    B2List:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/B2Object" }
...
    PublicId:
      type: string
      pattern: "^[A-Za-z0-9_-]{22}$"
      description: Deterministically encrypted reference; internal numeric key is never exposed.
    BookingReference:
      type: string
      pattern: "^LG-[A-Za-z0-9_-]{16}$"
      description: Opaque random reference used in customer-facing booking URLs.
...
    ApiError:
      type: object
      required: [code, message, requestId]
      properties:
        code: { type: string }
        message: { type: string }
        details: {}
        requestId: { type: string }
...